fix(resize): 리사이즈 에러 픽스
- 더미데이터 데이터 불러우는 위치 수정
- 헤더 사이즈 모바일도 80px로 수정

Resolve: #63
