+++ src/app/_components/Header/HeaderNav/HeaderNav.tsx
@@ -7,6 +7,8 @@ import { UserRole } from '@/_types/Header/Header.type';
 import { ReactNode, useEffect, useState } from 'react';
 import { HeaderOverlay, MenuButton } from './ETC';
 import useResize from '@/_hooks/nav/useResize';
+import { Alarm } from '@/_types/Header/Alarm.type';
+import useAlarmStore from '@/_store/Header/useAlarmStore';
 
 interface Props {
   children?: ReactNode;
@@ -18,6 +20,31 @@ const HeaderNav = ({ children }: Props) => {
   const [isToggle, setIsToggle] = useState(false);
   const viewmode = useHeaderStore((store) => store.state.mode);
   const toggleNavMobile = (state: boolean) => setIsToggle(state);
+  const addAlarm = useAlarmStore((store) => store.actions.addAlarm);
+
+  const dumyData: Alarm[] = [
+    {
+      id: 1,
+      lectureId: 2,
+      title: '강의 시간 변경',
+      contents: `03월 06일 ‘코드 그 너머: 소프트웨어 개발의 미래’
+      강의가 14시에서 15시로 시간이 변경되었습니다.`,
+    },
+    {
+      id: 1,
+      lectureId: 2,
+      title: '강의 시간 변경',
+      contents: `03월 07일 ‘코드 그 너머: 소프트웨어 개발의 미래’
+      강의가 13시에서 14시로 시간이 변경되었습니다.`,
+    },
+    {
+      id: 1,
+      lectureId: 2,
+      title: '강의 시간 변경',
+      contents: `03월 06일 ‘코드 그 너머: 소프트웨어 개발의 미래’
+      강의가  room-a1 에서 room-a2로 변경되었습니다.`,
+    },
+  ];
 
   // 리사이즈 훅
   useResize();
@@ -42,6 +69,16 @@ const HeaderNav = ({ children }: Props) => {
       return 'viewer';
     });
   };
+
+  // SSE 여기에 이벤트 구독 필요
+  useEffect(() => {
+    localStorage.removeItem('alarm-storage');
+    dumyData.forEach((item, idx) => {
+      setTimeout(() => {
+        addAlarm(item);
+      }, idx * 5000);
+    });
+  }, []);
 
   /** 웹 모드 **/
   if (viewmode === 'web') {
